src/requirements.txt
asgiref==3.2.7
Babel==2.8.0
boto3==1.13.14
botocore==1.16.14
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-environ==0.4.5
django-extensions==2.2.9
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-storages==1.9.1
django-widget-tweaks==1.4.8
docutils==0.15.2
environ==1.0
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
mysqlclient==1.4.6
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
twilio==6.41.0
urllib3==1.25.9
whitenoise==5.1.0
